refactor: directly pass BackupSchedulerCfg to DataCfg
Unfortunately, we cannot have the configuration as part of the unified config module, as it forms a cyclic dependency. Instead, we pass it directly from the unified config to the datacfg during startup.

Force retention schedule configuration existence.
